FIX: bunch.mean/sigma return np.float64
- mean_x/sigma_x... returned a np.array of size 1. Inconsistent with
  numpy, which returned a single float64
--> Wrap the functions with a cast to np.float64

Was a problem when the statistics were written to a file using .format
..., the identifier for a float is not the same as for an array

+++ particles/particles.py
@@ -199,40 +199,40 @@ class Particles(Printing):
     # Statistics methods
 
     def mean_x(self):
-        return pm.mean(self.x)
+        return np.float(pm.mean(self.x))
 
     def mean_xp(self):
-        return pm.mean(self.xp)
+        return np.float(pm.mean(self.xp))
 
     def mean_y(self):
-        return pm.mean(self.y)
+        return np.float(pm.mean(self.y))
 
     def mean_yp(self):
-        return pm.mean(self.yp)
+        return np.float(pm.mean(self.yp))
 
     def mean_z(self):
-        return pm.mean(self.z)
+        return np.float(pm.mean(self.z))
 
     def mean_dp(self):
-        return pm.mean(self.dp)
+        return np.float(pm.mean(self.dp))
 
     def sigma_x(self):
-        return pm.std(self.x)
+        return np.float(pm.std(self.x))
 
     def sigma_y(self):
-        return pm.std(self.y)
+        return np.float(pm.std(self.y))
 
     def sigma_z(self):
-        return pm.std(self.z)
+        return np.float(pm.std(self.z))
 
     def sigma_xp(self):
-        return pm.std(self.xp)
+        return np.float(pm.std(self.xp))
 
     def sigma_yp(self):
-        return pm.std(self.yp)
+        return np.float(pm.std(self.yp))
 
     def sigma_dp(self):
-        return pm.std(self.dp)
+        return np.float(pm.std(self.dp))
 
     def effective_normalized_emittance_x(self):
         return pm.emittance(self.x, self.xp, None) * self.betagamma
